benerin sorting lagi, add meeting button description forum, dan add animation

+++ database/migrations/2024_10_03_085022_investment.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('business_id');
-            $table->decimal('amount', 15, 2);
+            $table->integer('amount');
             $table->timestamps();
 
             // Foreign keys
